added xp of users above and below to API REST response
